Busca implementada para vários campos. Closes #1

+++ src/services/AlunoService.ts
@@ -21,11 +21,11 @@ export default class AlunoService implements IReadOnlyService, IEditService {
       if (!req.body._id) {
         model = await model.save();
       } else {
-        model = await AlunoCRUDModel.findOneAndUpdate({ _id: req.body._id }, model, {new : true})
+        model = await AlunoCRUDModel.findOneAndUpdate({ _id: req.body._id }, model, { new: true })
       }
 
-      if(!model){
-        return res.status(400).json({message : "A alteração de usuário resultou em erro"} as Error)
+      if (!model) {
+        return res.status(400).json({ message: "A alteração de usuário resultou em erro" } as Error)
       }
       return res.status(200).json(model);
     } catch (error) {
@@ -38,7 +38,7 @@ export default class AlunoService implements IReadOnlyService, IEditService {
       return res.status(400);
     }
     try {
-      let q = await AlunoCRUDModel.findOneAndRemove({ _id: req.params.id });
+      let q = await AlunoCRUDModel.findOneAndRemove({ _id: req.params.id, usuario: getUserIdFromRequest(req) });
       return res.status(200).json(q);
     } catch (error) {
       return res.status(500).json({ error });
@@ -47,7 +47,30 @@ export default class AlunoService implements IReadOnlyService, IEditService {
 
   async findAll(req: express.Request, res: express.Response) {
     try {
-      let results = await AlunoCRUDModel.find()
+      let results = await AlunoCRUDModel.find({ usuario: getUserIdFromRequest(req) })
+      return res.status(200).json(results);
+    } catch (error) {
+      return res.status(500).json(error)
+    }
+  }
+
+  async find(req: express.Request, res: express.Response) {
+    try {
+      let query = AlunoCRUDModel.find({ usuario: getUserIdFromRequest(req) })
+
+      const obj = req.body;
+      const keys = Object.keys(obj);
+
+      keys.forEach(key => {
+        if (typeof obj[key] === "string") {
+          query.where(key, new RegExp(obj[key], 'i'));
+        } else {
+          query.where(key, obj[key])
+        }
+      });
+
+      const results = await query.exec();
+
       return res.status(200).json(results);
     } catch (error) {
       return res.status(500).json(error)
@@ -56,7 +79,7 @@ export default class AlunoService implements IReadOnlyService, IEditService {
 
   async findOne(req: express.Request, res: express.Response) {
     try {
-      let result = await AlunoCRUDModel.findOne({ _id: req.params.id })
+      let result = await AlunoCRUDModel.findOne({ _id: req.params.id, usuario: getUserIdFromRequest(req) })
 
       if (result) {
         return res.status(200).json(result)
